+++ test/OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests/PrometheusIntegrationTests.cs
@@ -3,13 +3,19 @@
 
 using System.Diagnostics;
 using System.Diagnostics.Metrics;
+using System.Net;
+using System.Net.Http.Json;
+using System.Text.Json;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.Server;
 using Microsoft.AspNetCore.Hosting.Server.Features;
+using Microsoft.AspNetCore.WebUtilities;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using OpenTelemetry.Exporter.Prometheus.Tests;
 using OpenTelemetry.Metrics;
+using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
 using Xunit;
 using Xunit.Abstractions;
@@ -17,8 +23,238 @@ using Xunit.Abstractions;
 namespace OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests;
 
 [Collection(PromToolCollection.Name)]
-public class PrometheusIntegrationTests(PromToolFixture fixture, ITestOutputHelper outputHelper)
+public class PrometheusIntegrationTests(PromToolFixture promtool, ITestOutputHelper outputHelper)
 {
+    [Fact]
+    public async Task Scrape_Endpoint_Returns_No_Content_If_Sdk_Disabled()
+    {
+        // Arrange
+        var builder = WebApplication.CreateBuilder();
+
+        builder.Configuration.AddInMemoryCollection(
+            [
+                KeyValuePair.Create<string, string?>("OTEL_SDK_DISABLED", "true"),
+            ]);
+
+        // Listen on any available port
+        builder.WebHost.UseUrls("http://127.0.0.1:0");
+
+        builder.Services
+            .AddOpenTelemetry()
+            .WithMetrics((builder) => builder.AddPrometheusExporter());
+
+        using var app = builder.Build();
+
+        app.MapPrometheusScrapingEndpoint();
+
+        await app.StartAsync();
+
+        var server = app.Services.GetRequiredService<IServer>();
+        var addresses = server.Features.Get<IServerAddressesFeature>();
+
+        var baseAddress = addresses!.Addresses
+            .Select((p) => new Uri(p))
+            .Last();
+
+        using var httpClient = new HttpClient();
+        using var response = await httpClient.GetAsync(new Uri(baseAddress, "metrics"));
+
+        // Assert
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        Assert.Null(response.Content.Headers.ContentType);
+
+        var content = await response.Content.ReadAsStringAsync();
+
+        Assert.Empty(content);
+    }
+
+    [EnabledOnDockerPlatformTheory(DockerPlatform.Linux)]
+    [InlineData("")]
+    [InlineData("OpenMetricsText0.0.1")]
+    [InlineData("OpenMetricsText1.0.0")]
+    [InlineData("PrometheusText0.0.4")]
+    [InlineData("PrometheusText1.0.0")]
+
+    public async Task Prometheus_Can_Scrape_Metrics(string scrapeProtocol) => await GenerateMetricsAsync(async (baseAddress) =>
+    {
+        // Arrange
+        var prometheus = new PrometheusFixture()
+        {
+            TargetPort = baseAddress.Port,
+        };
+
+        if (!string.IsNullOrEmpty(scrapeProtocol))
+        {
+            prometheus.ScrapeProtocols.Add(scrapeProtocol);
+        }
+
+        try
+        {
+            // Act
+            await prometheus.StartAsync();
+
+            var prometheusBaseAddress = prometheus.GetBaseAddress(9090);
+
+            using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(20));
+
+            await WaitForServiceDiscoveryAsync(prometheusBaseAddress, outputHelper, cts.Token);
+
+            IReadOnlyList<string> series = [];
+
+            // Assert
+            while (!cts.IsCancellationRequested)
+            {
+                series = await WaitForMetricsSeriesAsync(prometheusBaseAddress, outputHelper, cts.Token);
+
+                if (series.Contains("temperature_celsius"))
+                {
+                    break;
+                }
+            }
+
+            Assert.Contains("http_server_active_requests", series);
+            Assert.Contains("http_server_request_duration_seconds_bucket", series);
+            Assert.Contains("http_server_request_duration_seconds_count", series);
+            Assert.Contains("http_server_request_duration_seconds_sum", series);
+            Assert.Contains("kestrel_active_connections", series);
+            Assert.Contains("kestrel_connection_duration_seconds_bucket", series);
+            Assert.Contains("kestrel_connection_duration_seconds_count", series);
+            Assert.Contains("kestrel_connection_duration_seconds_sum", series);
+            Assert.Contains("processed_bytes_total", series);
+            Assert.Contains("queue_balance", series);
+            Assert.Contains("temperature_celsius", series);
+
+#if NET10_0_OR_GREATER
+            Assert.Contains("aspnetcore_memory_pool_allocated_bytes_total", series);
+#endif
+        }
+        finally
+        {
+            (var stdout, var stderr) = await prometheus.TypedContainer.GetLogsAsync();
+
+            outputHelper.WriteLine($"[prometheus] [stdout]: {stdout}");
+            outputHelper.WriteLine($"[prometheus] [stderr]: {stderr}");
+
+            await prometheus.DisposeAsync();
+        }
+
+        static async Task<IReadOnlyList<string>> WaitForMetricsSeriesAsync(
+            Uri baseAddress,
+            ITestOutputHelper outputHelper,
+            CancellationToken cancellationToken)
+        {
+            // See https://prometheus.io/docs/prometheus/latest/querying/api/#finding-series-by-label-matchers
+            var seriesUrl = QueryHelpers.AddQueryString(
+                "/api/v1/series",
+                [
+                    KeyValuePair.Create<string, string?>("limit", "0"),
+                    KeyValuePair.Create<string, string?>("match[]", "{job=\"prometheus-target\"}"),
+                ]);
+
+            var seriesUri = new Uri(seriesUrl, UriKind.Relative);
+
+            var frequency = TimeSpan.FromMilliseconds(250);
+
+            using var client = new HttpClient() { BaseAddress = baseAddress };
+
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                try
+                {
+                    using var metrics = await client.GetFromJsonAsync<JsonDocument>(seriesUri, cancellationToken);
+
+                    if (metrics!.RootElement.ValueKind is JsonValueKind.Object &&
+                        metrics.RootElement.TryGetProperty("status", out var status) &&
+                        status.GetString() == "success")
+                    {
+                        var data = metrics.RootElement.GetProperty("data");
+
+                        if (data.GetArrayLength() > 0)
+                        {
+                            var series = new HashSet<string>();
+
+                            foreach (var seriesElement in data.EnumerateArray())
+                            {
+                                if (seriesElement.ValueKind is JsonValueKind.Object &&
+                                    seriesElement.TryGetProperty("__name__", out var name))
+                                {
+                                    series.Add(name.GetString()!);
+                                }
+                            }
+
+                            return [.. series];
+                        }
+                    }
+                }
+                catch (Exception ex)
+                {
+                    outputHelper.WriteLine($"[prometheus] Exception while waiting for metric series: {ex}");
+
+                    try
+                    {
+                        await Task.Delay(frequency, cancellationToken);
+                    }
+                    catch (TaskCanceledException)
+                    {
+                        break;
+                    }
+                }
+            }
+
+            Assert.Fail($"Timed out waiting for metric series.");
+            return [];
+        }
+
+        static async Task WaitForServiceDiscoveryAsync(
+            Uri baseAddress,
+            ITestOutputHelper outputHelper,
+            CancellationToken cancellationToken)
+        {
+            // See https://prometheus.io/docs/prometheus/latest/querying/api/#targets
+            using var client = new HttpClient() { BaseAddress = baseAddress };
+            var targetsUri = new Uri("/api/v1/targets", UriKind.Relative);
+
+            var frequency = TimeSpan.FromMilliseconds(250);
+
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                try
+                {
+                    using var targets = await client.GetFromJsonAsync<JsonDocument>(targetsUri, cancellationToken);
+
+                    if (targets!.RootElement.ValueKind is JsonValueKind.Object &&
+                        targets.RootElement.TryGetProperty("status", out var status) &&
+                        status.GetString() == "success")
+                    {
+                        var activeTargets = targets.RootElement
+                            .GetProperty("data")
+                            .GetProperty("activeTargets");
+
+                        if (activeTargets.GetArrayLength() > 0)
+                        {
+                            return;
+                        }
+                    }
+                }
+                catch (Exception ex)
+                {
+                    outputHelper.WriteLine($"[prometheus] Exception while waiting for service discovery: {ex}");
+
+                    try
+                    {
+                        await Task.Delay(frequency, cancellationToken);
+                    }
+                    catch (TaskCanceledException)
+                    {
+                        break;
+                    }
+                }
+            }
+
+            Assert.Fail($"Timed out waiting for service discovery active targets.");
+        }
+    });
+
     [EnabledOnDockerPlatformTheory(DockerPlatform.Linux)]
     [InlineData("")]
     [InlineData("text/plain")]
@@ -28,8 +264,32 @@ public class PrometheusIntegrationTests(PromToolFixture fixture, ITestOutputHelp
     [InlineData("application/openmetrics-text;version=0.0.4")]
     [InlineData("application/openmetrics-text;version=1.0.0", Skip = "https://github.com/prometheus/prometheus/issues/8932")]
     [InlineData("application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.5,application/openmetrics-text;version=0.0.1;q=0.4,text/plain;version=1.0.0;escaping=allow-utf-8;q=0.3,text/plain;version=0.0.4;q=0.2,/;q=0.1", Skip = "https://github.com/prometheus/prometheus/issues/8932")]
+    public async Task Promtool_Considers_Scrape_Response_Valid(string accept) => await GenerateMetricsAsync(async (baseAddress) =>
+    {
+        // Act
+        var actual = await promtool.CheckMetricsAsync(new(baseAddress, "metrics"), accept);
 
-    public async Task Can_Scrape_Prometheus(string accept)
+        outputHelper.WriteLine($"[promtool] ExitCode: {actual.ExitCode}");
+        outputHelper.WriteLine("[promtool] stdout:");
+        outputHelper.WriteLine(string.Empty);
+        outputHelper.WriteLine(actual.Stdout);
+
+        if (!string.IsNullOrEmpty(actual.Stderr))
+        {
+            outputHelper.WriteLine(string.Empty);
+            outputHelper.WriteLine("[promtool] stderr:");
+            outputHelper.WriteLine(string.Empty);
+            outputHelper.WriteLine(actual.Stderr);
+        }
+
+        // Assert
+        Assert.Equal(0, actual.ExitCode);
+        Assert.NotEmpty(actual.Stdout);
+        Assert.Empty(actual.Stderr);
+    });
+
+    private static async Task GenerateMetricsAsync(
+        Func<Uri, Task> actAndAssert)
     {
         // Arrange
         const string meterName = "prometheus.integration.tests";
@@ -73,6 +333,7 @@ public class PrometheusIntegrationTests(PromToolFixture fixture, ITestOutputHelp
 
         builder.Services
             .AddOpenTelemetry()
+            .ConfigureResource((builder) => builder.AddService("my-service", "my-namespace", "1.2.3"))
             .WithMetrics((builder) =>
             {
                 builder.AddAspNetCoreInstrumentation()
@@ -152,26 +413,8 @@ public class PrometheusIntegrationTests(PromToolFixture fixture, ITestOutputHelp
 
         activity.Stop();
 
-        // Act
-        var actual = await fixture.CheckMetricsAsync(new(baseAddress, "metrics"), accept);
-
-        outputHelper.WriteLine($"[promtool] ExitCode: {actual.ExitCode}");
-        outputHelper.WriteLine("[promtool] stdout:");
-        outputHelper.WriteLine(string.Empty);
-        outputHelper.WriteLine(actual.Stdout);
-
-        if (!string.IsNullOrEmpty(actual.Stderr))
-        {
-            outputHelper.WriteLine(string.Empty);
-            outputHelper.WriteLine("[promtool] stderr:");
-            outputHelper.WriteLine(string.Empty);
-            outputHelper.WriteLine(actual.Stderr);
-        }
-
-        // Assert
-        Assert.Equal(0, actual.ExitCode);
-        Assert.NotEmpty(actual.Stdout);
-        Assert.Empty(actual.Stderr);
+        // Act and Assert
+        await actAndAssert(baseAddress);
     }
 
     private static void WaitForNextExemplarTimestamp()